VisiChunk: Jetzt gibt es fast das selbe aus - WTF?
--- alf.html    2013-05-28 11:18:53.082979715 +0200
+++ alf4cpp.html        2013-07-01 06:53:16.975845598 +0200
@@ -116,7 +116,7 @@
 New file   : /home/raimi/VMs/alf4devel/alf4devel-disk1.vmdk (8149467136 b)
 Chunk size : 1048576
 Common bytes   : 7949254656    97.6% of 8148811776
-Changed bytes  : 10610839      0.1% of 8148811776
+Changed bytes  : 10610838      0.1% of 8148811776
 Equal Chunks   : 7581  97.5%
 Changed Chunks : 191   2.5%
 </pre>

+++ src/VisiChunk.cpp
@@ -206,7 +206,7 @@ void run()
         chgtotal += cmpdiff;
         numChange++;
         // Anderer Block
-        int cl = ceil(cmpdiff * 100/ cmpsize);
+        int cl = ceil((double) cmpdiff * 100 / (double) cmpsize);
         if (html) {
           printf ("<div class=\"c c%i\"></div>", cl);
         } else {
